pyisobmff: add support for meta box issue
The MetaBox is a FullBox item, i.e., it uses the 4 bytes
just following the "meta" box type to include version
and flags. The MetaBox is typically followed by an "hdlr"
box, so the distance between both box types is in this
case 8 bytes.

Good example:
```
$ xxd videolan/samples/mov/HEIF/Alchimiste.HEIC
...
00000010: 6d69 6631 6865 6963 0000 0f73 6d65 7461  mif1heic...smeta
00000020: 0000 0000 0000 0022 6864 6c72 0000 0000  ......."hdlr....
...
```

Note the 8 bytes between "meta" and "hdlr"

There is a relatively common issue on isobmff streams where
the MetaBox is actually a Box, not a FullBox. This means the
parser reads the FullBox flags and version fields incorrectly
(from the hdlr box size field), and then reads the "hdlr" box
type as a length. Parsing crashes then.

We will peek at the next 4x bytes before reading the very first
box in the MetaBox. We detect the issue if we find it is "hdlr".
In that case we need to:
(1) reset the flags and version fields.
(2) seek back the wrongly-read 4x bytes.

Broken example:
```
$ xxd videolan/samples/mov/editlist/20210910_114302.mp4
...
05a73ba0: 7400 0000 0000 0000 0000 b16d 6574 6100  t..........meta.
05a73bb0: 0000 2168 646c 7200 0000 0000 0000 006d  ..!hdlr........m
...
```

Note the 4 bytes between "meta" and "hdlr"